requirements.txt
asgiref==3.5.2
mysqlclient==2.1.0
PyMySQL==1.0.2
pycparser==2.21
PyJWT==2.4.0
Django==4.0.4
django-allauth==0.44.0
django-auth==0.1.9
django-crispy-forms==1.12.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.2.0
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
djangorestframework==3.13.1
